Collapse `ReachingDefFlowGraph` into `FlowGraph`
